+++ src/snowcli/app/main_typer.py
@@ -1,7 +1,7 @@
 import sys
 
 import click
-from snowcli.cli.common.cli_global_context import cli_context, cli_context_manager
+from snowcli.cli.common.cli_global_context import cli_context
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, DebugOption
 from typer import Typer
