+++ tests_e2e/test_installation.py
@@ -51,37 +51,3 @@ def test_command_from_external_plugin(snowcli, test_root_path, snapshot):
         encoding="utf-8",
     )
     snapshot.assert_match(output)
-
-
-@pytest.mark.e2e
-def test_error_traceback_disabled_without_debug(snowcli, test_root_path):
-    traceback_msg = "Traceback (most recent call last)"
-
-    result = subprocess.run(
-        [
-            snowcli,
-            "--config-file",
-            test_root_path / "config" / "malformatted_config.toml",
-            "sql",
-            "-q",
-            "select 'Hello there'",
-        ],
-        capture_output=True,
-        text=True,
-    )
-    assert result.stderr == "" and not traceback_msg in result.stdout
-
-    result_debug = subprocess.run(
-        [
-            snowcli,
-            "--config-file",
-            test_root_path / "config" / "malformatted_config.toml",
-            "sql",
-            "-q",
-            "select 'Hello there'",
-            "--debug",
-        ],
-        capture_output=True,
-        text=True,
-    )
-    assert traceback_msg in result_debug.stderr
